vlm_fine_tuned/requirements.txt
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.27.0
peft>=0.10.0
trl>=0.8.0
datasets>=2.18.0
pillow>=10.0.0
bitsandbytes>=0.43.0
scipy
num2words>=0.5.0

